+++ plugins/postcss-base-plugin/test/_import.mjs
@@ -1,6 +1,10 @@
-import assert from 'assert';
+import assert from 'node:assert/strict';
+import test from 'node:test';
 import plugin from '@csstools/postcss-base-plugin';
-plugin();
 
-assert.ok(plugin.postcss, 'should have "postcss flag"');
-assert.equal(typeof plugin, 'function', 'should return a function');
+test('import', () => {
+	plugin();
+	assert.ok(plugin.postcss, 'should have "postcss flag"');
+	assert.equal(typeof plugin, 'function', 'should return a function');
+});
+