+++ plugins/postcss-focus-visible/test/_browser.mjs
@@ -2,38 +2,47 @@
 import puppeteer from 'puppeteer';
 import http from 'http';
 import { promises as fsp } from 'fs';
+import test from 'node:test';
+import process from 'node:process';
 
-(async () => {
-	const requestListener = async function (req, res) {
+const requestListener = async function (req, res) {
 
-		const parsedUrl = new URL(req.url, 'http://localhost:8080');
-		const pathname = parsedUrl.pathname;
+	const parsedUrl = new URL(req.url, 'http://localhost:8080');
+	const pathname = parsedUrl.pathname;
 
-		switch (pathname) {
-			case '':
-			case '/':
-				res.setHeader('Content-type', 'text/html');
-				res.writeHead(200);
-				res.end(await fsp.readFile('test/_browser.html', 'utf8'));
-				break;
-			case '/test/browser.expect.css':
-				res.setHeader('Content-type', 'text/css');
-				res.writeHead(200);
-				res.end(await fsp.readFile('test/browser.expect.css', 'utf8'));
-				break;
-			default:
-				res.setHeader('Content-type', 'text/plain');
-				res.writeHead(404);
-				res.end('Not found');
-				break;
-		}
-	};
+	switch (pathname) {
+		case '':
+		case '/':
+			res.setHeader('Content-type', 'text/html');
+			res.writeHead(200);
+			res.end(await fsp.readFile('test/_browser.html', 'utf8'));
+			break;
+		case '/test/browser.expect.css':
+			res.setHeader('Content-type', 'text/css');
+			res.writeHead(200);
+			res.end(await fsp.readFile('test/browser.expect.css', 'utf8'));
+			break;
+		default:
+			res.setHeader('Content-type', 'text/plain');
+			res.writeHead(404);
+			res.end('Not found');
+			break;
+	}
+};
 
-	// Use different servers for HTML/CSS/JS to trigger CORS
+function startServers() {
 	const server = http.createServer(requestListener);
 	server.listen(8080);
 
-	if (!process.env.DEBUG) {
+	return () => {
+		server.close();
+	};
+}
+
+if (!process.env.DEBUG) {
+	test('browser', { skip: process.env.GITHUB_ACTIONS && !process.env.BROWSER_TESTS }, async () => {
+		const cleanup = startServers();
+
 		const browser = await puppeteer.launch({
 			headless: 'new',
 		});
@@ -90,8 +99,10 @@ import { promises as fsp } from 'fs';
 
 		await browser.close();
 
-		await server.close();
-	} else {
-		console.log('visit : http://localhost:8080');
-	}
-})();
+		await cleanup();
+	});
+} else {
+	startServers();
+
+	console.log('visit : http://localhost:8080');
+}