+++ plugins/postcss-progressive-custom-properties/.tape.mjs
@@ -1,8 +1,0 @@
-import { postcssTape } from '@csstools/postcss-tape';
-import plugin from '@csstools/postcss-progressive-custom-properties';
-
-postcssTape(plugin)({
-	basic: {
-		message: "supports basic usage",
-	},
-});