+++ plugins/postcss-progressive-custom-properties/test/_tape.mjs
@@ -1,0 +1,8 @@
+import { postcssTape } from '@csstools/postcss-tape';
+import plugin from '@csstools/postcss-progressive-custom-properties';
+
+postcssTape(plugin)({
+	basic: {
+		message: 'supports basic usage',
+	},
+});