+++ rollup/transforms/node-coverage-disable.mjs
@@ -1,0 +1,8 @@
+export function nodeCoverageDisable () {
+	return {
+		name: 'node-coverage-disable',
+		renderChunk (code) {
+			return `/* node:coverage disable */\n${code}\n/* node:coverage enable */\n`;
+		},
+	};
+}