+++ packages/postcss-tape/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changes to PostCSS Tape
 
+## Unreleased (major)
+
+- Use `node:test` instead of custom logic.
+
 ### 3.0.2
 
 _December 15, 2023_